requirements.txt
absl-py==2.1.0
aenum==3.1.15
affine==2.4.0
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
albucore==0.0.15
albumentations==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs @ file:///croot/attrs_1695717823297/work
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
blosc2==2.7.1
boto3==1.34.131
botocore==1.34.131
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.2
contourpy==1.2.1
coverage==7.4.4
croniter==1.4.1
cryptacular==1.6.2
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.1
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
dnspython==2.6.1
docker==6.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
editor==1.6.6
efficientnet-pytorch==0.7.1
einops==0.8.1
email_validator==2.2.0
eval_type_backport==0.2.0
evdev==1.7.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
fast-pytorch-kmeans==0.2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock @ file:///croot/filelock_1700591183607/work
fiona==1.9.6
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.3.1
gdown==5.2.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
googledrivedownloader==0.4
GPUtil==1.4.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.1
hupper==1.12.1
hydra-core==1.3.2
idna @ file:///croot/idna_1714398848350/work
imageio==2.34.2
imgaug==0.4.0
importlib_resources==6.4.0
iniconfig==2.0.0
inquirer==3.3.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
jmespath==1.0.1
joblib==1.4.2
json5==0.9.24
jsonargparse==4.31.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter @ file:///croot/jupyter_1707947101020/work
jupyter-console @ file:///croot/jupyter_console_1679999630278/work
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightly==1.5.8
lightly-utils==0.0.2
lightning @ file:///home/conda/feedstock_root/build_artifacts/lightning_1720295862678/work
lightning-api-access==0.0.5
lightning-cloud==0.5.70
lightning-fabric==2.3.3
lightning-utilities @ file:///home/conda/feedstock_root/build_artifacts/lightning-utilities_1719444219337/work
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpld3==0.5.10
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.8
multidict==6.0.5
munch==4.0.0
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
ndindex==1.9.2
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1720002482208/work
notebook==7.1.2
notebook_shim==0.2.4
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1714398848491/work
pipreqs==0.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
pluggy==1.5.0
ply==3.11
pretrainedmodels==0.7.4
prettytable==3.9.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.8
ptflops==0.7.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic_core==2.20.1
Pygments==2.17.2
PyJWT==2.8.0
pynput==1.7.7
pynvml==11.5.0
pyparsing==3.1.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.4
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-xlib==0.33
python3-openid==3.2.0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.3.0
pytorch-model-summary==0.1.2
pytorch-ranger==0.1.1
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==25.1.2
qtconsole @ file:///croot/qtconsole_1709231153903/work
QtPy @ file:///croot/qtpy_1700144840038/work
qudida==0.0.4
rasterio==1.3.10
readchar==4.1.0
redis==5.0.7
referencing==0.34.0
regex==2023.12.25
repoze.sendmail==4.4.1
requests @ file:///croot/requests_1716902831423/work
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
Rtree==1.2.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
runs==1.2.2
s3fs==2023.12.2
s3transfer==0.10.2
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.3
sentry-sdk==2.10.0
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
shtab==1.7.1
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
sympy @ file:///croot/sympy_1701397643339/work
tables==3.10.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==1.0.15
tinycss2==1.2.1
tokenizers==0.20.1
tomli==2.0.1
torch==2.1.0
torch-optimizer==0.3.0
torchaudio==2.2.2+cu121
torchgeo==0.5.2
torchinfo==1.8.0
torchmetrics==1.2.0
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.16.0
tornado==6.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1714854870413/work
traitlets==5.14.2
transaction==4.0
transformers==4.45.2
translationstring==1.4
triton==2.1.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typeshed_client==2.5.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tyro==0.8.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1718912636303/work
uvicorn==0.30.1
uvloop==0.19.0
velruse==1.1.1
venusian==3.1.0
ViTables==3.0.3
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.8
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xformers==0.0.22.post7
xmod==1.8.1
yarg==0.1.9
yarl==1.9.4
zope.deprecation==5.0
zope.interface==7.0.1
zope.sqlalchemy==3.1
